Attach SecretTypeDockerConfigJson secret type for deployments
The SecretTypeDockerConfigJson is now used as of Kubernetes 1.9
to store image pull secrets for fetching images from private
repositories. [0] This change allows SecretTypeDockerConfigJson
as a parallel option to the existing option SecretTypeDockercfg.

[0] https://kubernetes.io/docs/tasks/configure-pod-container/pull-image-private-registry/
