+++ handlers/deploy.go
@@ -31,9 +31,6 @@ const watchdogPort = 8080
 // initialReplicasCount how many replicas to start of creating for a function
 const initialReplicasCount = 1
 
-// DefaultFunctionNamespace define default work namespace
-const DefaultFunctionNamespace string = "default"
-
 // ValidateDeployRequest validates that the service name is valid for Kubernetes
 func ValidateDeployRequest(request *requests.CreateFunctionRequest) error {
 	// Regex for RFC-1123 validation:
@@ -217,89 +214,6 @@ func makeDeploymentSpec(request requests.CreateFunctionRequest, existingSecrets 
 	}
 
 	return deploymentSpec, nil
-}
-
-// UpdateSecrets will update the Deployment spec to include secrets that have beenb deployed
-// in the kubernetes cluster.  For each requested secret, we inspect the type and add it to the
-// deployment spec as appropriat: secrets with type `SecretTypeDockercfg` are added as ImagePullSecrets
-// all other secrets are mounted as files in the deployments containers.
-func UpdateSecrets(request requests.CreateFunctionRequest, deployment *v1beta1.Deployment, existingSecrets map[string]*apiv1.Secret) error {
-	// Add / reference pre-existing secrets within Kubernetes
-	secretVolumeProjections := []apiv1.VolumeProjection{}
-	for _, secretName := range request.Secrets {
-		deployedSecret, ok := existingSecrets[secretName]
-		if !ok {
-			return fmt.Errorf("Required secret '%s' was not found in the cluster", secretName)
-		}
-
-		if deployedSecret.Type == apiv1.SecretTypeDockercfg {
-			deployment.Spec.Template.Spec.ImagePullSecrets = append(
-				deployment.Spec.Template.Spec.ImagePullSecrets,
-				apiv1.LocalObjectReference{
-					Name: secretName,
-				},
-			)
-		} else {
-			// projectSecrets.VolumeSource.Sources = newProjections
-
-			projectedPaths := []apiv1.KeyToPath{}
-			for secretKey := range deployedSecret.Data {
-				projectedPaths = append(projectedPaths, apiv1.KeyToPath{Key: secretKey, Path: secretKey})
-			}
-
-			projection := &apiv1.SecretProjection{Items: projectedPaths}
-			projection.Name = secretName
-			secretProjection := apiv1.VolumeProjection{
-				Secret: projection,
-			}
-			secretVolumeProjections = append(secretVolumeProjections, secretProjection)
-
-		}
-	}
-
-	if len(secretVolumeProjections) > 0 {
-		volumeName := fmt.Sprintf("%s-projected-secrets", request.Service)
-		projectedSecrets := apiv1.Volume{
-			Name: volumeName,
-			VolumeSource: apiv1.VolumeSource{
-				Projected: &apiv1.ProjectedVolumeSource{
-					Sources: secretVolumeProjections,
-				},
-			},
-		}
-		deployment.Spec.Template.Spec.Volumes = append(deployment.Spec.Template.Spec.Volumes, projectedSecrets)
-
-		// add mount secret as a file
-		updatedContainers := []apiv1.Container{}
-		for _, container := range deployment.Spec.Template.Spec.Containers {
-			mount := apiv1.VolumeMount{
-				Name:      volumeName,
-				ReadOnly:  true,
-				MountPath: "/run/secrets",
-			}
-			container.VolumeMounts = append(container.VolumeMounts, mount)
-			updatedContainers = append(updatedContainers, container)
-		}
-
-		deployment.Spec.Template.Spec.Containers = updatedContainers
-	}
-
-	return nil
-}
-
-// getSecrets queries Kubernetes for a list of secrets by name in the given k8s namespace.
-func getSecrets(clientset *kubernetes.Clientset, namespace string, secretNames []string) (map[string]*apiv1.Secret, error) {
-	secrets := map[string]*apiv1.Secret{}
-
-	for _, secretName := range secretNames {
-		secret, err := clientset.Core().Secrets(namespace).Get(secretName, metav1.GetOptions{})
-		if err != nil {
-			return secrets, err
-		}
-		secrets[secretName] = secret
-	}
-
-	return secrets, nil
 }
 
 func makeServiceSpec(request requests.CreateFunctionRequest) *v1.Service {
